+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Replace checkout line items amount `code` field with encrypted amount field.
+- Rename checkout lines items `id` field to `product_id`.
 
 ### Removed
 - Checkout line items data validation.